Don't add genes that are expected to be renamed in the COVERAGE_DICTIONARY
This might have unintended consequences, but as far as I can see there
isn't a reason to add these in the COVERAGE_DICTIONARY. This fixes an issue
where both mmpR5 and Rv0678 were being counted when calculating the
`percentage_lims_genes_above` >= `MIN_LOCUS_PERCENTAGE`. This has
implications for which lineage designation is assigned.